+++ R/bbk-package.R
@@ -2,6 +2,7 @@
 #' @import checkmate
 #' @import data.table
 #' @importFrom httr2 req_body_json
+#' @importFrom httr2 req_cache
 #' @importFrom httr2 req_error
 #' @importFrom httr2 req_headers
 #' @importFrom httr2 req_perform